+++ challenges/backend/src/app/main.ts
@@ -8,6 +8,8 @@ import { Config } from "./services/Config/classes/Config";
 import { ICalculator } from "./services/CarOnSaleClient/interface/ICalculator";
 import { Calculator } from "./services/CarOnSaleClient/classes/Calculator";
 import axios, { AxiosInstance } from "axios";
+import { ICarOnSaleClient } from "./services/CarOnSaleClient/interface/ICarOnSaleClient";
+import { CarOnSale } from "./services/CarOnSaleClient/classes/CarOnSale";
 
 /*
  * Create the DI container.
@@ -25,6 +27,7 @@ container.bind<ILogger>(DependencyIdentifier.LOGGER).to(Logger);
 container.bind<IConfig>(DependencyIdentifier.CONFIG).to(Config);
 container.bind<ICalculator>(DependencyIdentifier.CALCULATOR).to(Calculator);
 container.bind<AxiosInstance>(DependencyIdentifier.CLIENT).toConstantValue(axiosInstance);
+container.bind<ICarOnSaleClient>(DependencyIdentifier.CAR_ON_SALE).to(CarOnSale);
 
 
 /*
